Refactoring of the code
